apis_py/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
pydantic-settings
python-jose
PyJWT
flagger
python-multipart